+++ agent/api/container.go
@@ -123,6 +123,10 @@ type Container struct {
 	// handled properly so that the state storage continues to work.
 	SentStatusUnsafe ContainerStatus `json:"SentStatus"`
 
+	// MetadataFileUpdated is set to true when we have completed updating the
+	// metadata file
+	MetadataFileUpdated bool `json:"metadataFileUpdated"`
+
 	knownExitCode     *int
 	KnownPortBindings []PortBinding
 
@@ -161,19 +165,6 @@ func NewContainerWithSteadyState(steadyState ContainerStatus) *Container {
 	return &Container{
 		SteadyStateStatusUnsafe: &steadyStateStatus,
 	}
-}
-
-// Overriden applies the overridden command and returns the resulting
-// container object
-func (c *Container) Overridden() *Container {
-	result := *c
-
-	// We only support Command overrides at the moment
-	if result.Overrides.Command != nil {
-		result.Command = *c.Overrides.Command
-	}
-
-	return &result
 }
 
 // KnownTerminal returns true if the container's known status is STOPPED
@@ -316,4 +307,21 @@ func (c *Container) IsInternal() bool {
 // or RESOURCES_PROVISIONED. It returns false otherwise
 func (c *Container) IsRunning() bool {
 	return c.GetKnownStatus().IsRunning()
+}
+
+// IsMetadataFileUpdated returns true if the metadata file has been once the
+// metadata file is ready and will no longer change
+func (c *Container) IsMetadataFileUpdated() bool {
+	c.lock.RLock()
+	defer c.lock.RUnlock()
+
+	return c.MetadataFileUpdated
+}
+
+// SetMetadataFileUpdated sets the container's MetadataFileUpdated status to true
+func (c *Container) SetMetadataFileUpdated() {
+	c.lock.Lock()
+	defer c.lock.Unlock()
+
+	c.MetadataFileUpdated = true
 }